+++ backend/src/users/users.controller.ts
@@ -17,7 +17,7 @@ import {
   UploadedFile,
 } from '@nestjs/common';
 import { Cache } from 'cache-manager';
-import { Prisma, Role } from '@prisma/client';
+import { Prisma } from '@prisma/client';
 import { SessionContainer } from 'supertokens-node/recipe/session';
 import ses from 'supertokens-node/recipe/session';
 
@@ -28,20 +28,16 @@ import { JoiValidationPipe } from '../Pipes/JoiValidationPipe';
 import { FilesPipe } from '../Pipes/FilesPipe';
 import { UsersPipe } from '../Pipes/UsersPipe';
 import { allContent } from '../constants/allCustomsHttpMessages';
-import { UserDto, SortType, MembersDto } from '../DTOs/user.dto';
+import { UserDto, SortType } from '../DTOs/user.dto';
 import { QueryDto } from '../DTOs/query.dto';
 
-import { PrismaService } from '../prisma/prisma.service';
 import { UsersService } from './users.service';
-import { UsersGroupsService } from '../users-groups/users-groups.service';
 
 @Controller('users')
 @UseInterceptors(CacheInterceptor)
 export class UsersController {
   constructor(
-    private prisma: PrismaService,
     private readonly usersService: UsersService,
-    private readonly usersGroupsService: UsersGroupsService,
     @Inject(CACHE_MANAGER) private cacheManager: Cache,
   ) {}
 
@@ -66,9 +62,6 @@ export class UsersController {
         where: whereElements,
       });
 
-      const firstNextData: UserDto[] = [];
-      const nextData: UserDto[] = [];
-
       if (!!cursor) {
         const nextResults = await this.usersService.findAllUsers({
           take: parseInt(limit),
@@ -81,111 +74,14 @@ export class UsersController {
         });
 
         if (nextResults.length > 0) {
-          if (firstNextData.length === 0) {
-            firstNextData.concat(firstResults, nextResults);
-            await this.cacheManager.set('users', firstNextData);
-            return firstNextData;
-          }
-
-          if (nextData.length === 0) {
-            nextData.concat(firstNextData, nextResults);
-            await this.cacheManager.set('users', nextData);
-            return nextData;
-          }
-
-          nextData.concat(nextResults);
-          await this.cacheManager.set('users', nextData);
-          return nextData;
+          await this.cacheManager.set('users', nextResults);
+          return nextResults;
         } else {
           return allContent;
         }
       } else {
         await this.cacheManager.set('users', firstResults);
         return firstResults;
-      }
-    }
-  }
-
-  @Get('members')
-  @UseGuards(new AuthGuard())
-  async getUsersFromGroupsWithRole(
-    @Query('queryData')
-    queryData: QueryDto & { name: string; role: Role },
-  ) {
-    const { orderBy, limit, cursor, name, role } = queryData;
-
-    const { roleId } = await this.prisma.roles.findUnique({
-      where: { type: role },
-      select: {
-        roleId: true,
-      },
-    });
-
-    const { order }: SortType = await queriesTransformation(false, orderBy);
-
-    const firstIdResults =
-      await this.usersGroupsService.getTheGroupOfGivenUserAboutGivenRole({
-        orderBy: order,
-        take: parseInt(limit),
-        name,
-        roleId,
-      });
-
-    if (role === Role.ADMIN) {
-      const { pseudonym, profilePhoto } = await this.usersService.findUser({
-        id: firstIdResults[0].userId,
-      });
-      return {
-        pseudonym,
-        profilePhoto,
-      };
-    } else {
-      const firstData: MembersDto[] = [];
-
-      for (const first of firstIdResults) {
-        const user: UserDto = await this.usersService.findUser({
-          id: first.userId,
-        });
-
-        firstData.push({
-          usersGroupsId: first.usersGroupsId,
-          pseudonym: user.pseudonym,
-          profilePhoto: user.profilePhoto,
-        });
-      }
-
-      if (!!cursor) {
-        const nextData: MembersDto[] = [];
-
-        const nextIdResults =
-          await this.usersGroupsService.getTheGroupOfGivenUserAboutGivenRole({
-            orderBy: order,
-            take: parseInt(limit),
-            skip: 1,
-            cursor: {
-              usersGroupsId: cursor,
-            },
-            name,
-            roleId,
-          });
-
-        if (nextIdResults.length > 0) {
-          for (const next of nextIdResults) {
-            const user: UserDto = await this.usersService.findUser({
-              id: next.userId,
-            });
-            nextData.push({
-              usersGroupsId: next.usersGroupsId,
-              pseudonym: user.pseudonym,
-              profilePhoto: user.profilePhoto,
-            });
-          }
-
-          await this.cacheManager.set('usersGroups', nextData);
-          return nextData;
-        }
-      } else {
-        return firstData;
       }
     }
   }
@@ -257,11 +153,7 @@ export class UsersController {
     @Session() session: SessionContainer,
     @Param('pseudonym') pseudonym: string,
   ) {
-    return this.usersService.deleteUser(
-      { pseudonym },
-      session.getUserId(),
-      ses,
-      session,
-    );
+    const userId = session.getUserId();
+    return this.usersService.deleteUser({ pseudonym }, userId, ses, session);
   }
 }
